Main 2023

+++ install_dotfiles.sh
@@ -1,4 +1,4 @@
-# /bin/bash
+#!/bin/bash
 
 if [[ $UID != 0 ]]; then
     echo "Please run this script with sudo:"
@@ -6,9 +6,20 @@ if [[ $UID != 0 ]]; then
     exit 1
 fi
 
-echo "This script will copy system configuration files and install dotfiles  in both non-root and root users"
-echo "Type the non-root user"
-read non_root
+non_root=$1
+
+if [[ -z "$non_root" ]]; then
+    echo "Please provide a non-root user as an argument:"
+    echo "sudo $0 <non-root user>"
+    exit 1
+fi
+
+if id "$non_root" >/dev/null 2>&1; then
+    echo "User $non_root exists"
+else
+    echo "User $non_root does not exist"
+    exit 1
+fi
 
 echo "Linking non-root files"
 ln -s /home/$non_root/git/fresh-linux/dotfiles/server/not-root/.bash_aliases /home/$non_root/.bash_aliases
@@ -20,20 +31,21 @@ ln -s /home/$non_root/git/fresh-linux/dotfiles/server/not-root/.nanorc /home/$no
 
 
 echo "Linking root files"
-sudo ln -s /home/$non_root/git/fresh-linux/dotfiles/server/not-root/.bash_aliases /root/.bash_aliases
-sudo rm -f /root/.bashrc
-sudo ln -s /home/$non_root/git/fresh-linux/dotfiles/server/root/.bashrc /root/.bashrc
-sudo rm -f /root/.bash_profile
-sudo ln -s /home/$non_root/.bash_profile /root/.bash_profile
+if [ -f "/home/$non_root/git/fresh-linux/dotfiles/server/root/.bash_aliases" ]; then
+    sudo ln -s /home/$non_root/git/fresh-linux/dotfiles/server/root/.bash_aliases /root/.bash_aliases
+fi
+if [ -f "/home/$non_root/git/fresh-linux/dotfiles/server/root/.bashrc" ]; then
+    sudo rm -f /root/.bashrc
+    sudo ln -s /home/$non_root/git/fresh-linux/dotfiles/server/root/.bashrc /root/.bashrc
+fi
+if [ -f "/home/$non_root/git/fresh-linux/dotfiles/server/root/.bash_profile" ]; then
+    sudo rm -f /root/.bash_profile
+    sudo ln -s /home/$non_root/git/fresh-linux/dotfiles/server/root/.bash_profile /root/.bash_profile
+fi
 
 echo "Copying system configuration files"
-sudo ln -s /root/config/etc/asound.conf /etc/asound.conf
-sudo ln -s /root/config/etc/mpd.conf /etc/mpd.conf
-sudo mkdir /etc/spotifyd
-sudo ln -s /root/config/etc/spotifyd/* /etc/spotifyd
-sudo ln -s /root/config/etc/systemd/system/* /etc/systemd/system
-sudo ln -s /root/config/etc/logrotate.d/* /etc/logrotate.d
-sudo ln -s /root/config/var/spool/cron/crontabs/root /var/spool/cron/crontabs/
-sudo ln -s /root/config/etc/wireguard/wg0.conf /etc/wireguard
-mkdir /home/$non_root/.config/mpd
-sudo ln -s /root/config/home/user/.config/mpd/mpd.conf /home/feder/.config/mpd/
+if [ -f "/root/config/etc/asound.conf" ]; then
+    sudo ln -s /root/config/etc/asound.conf /etc/asound.conf
+fi
+if [ -f "/root/config/etc/mpd.conf" ]; then
+    sudo ln -s /root/config/etc/mpd.conf /etc
